ECC-1352: CCSDS with multi-threaded tests

+++ tests/grib_encode_pthreads.sh
@@ -21,14 +21,18 @@ done
 
 rm -f temp.$label.out_*.grib
 
+if [ $HAVE_AEC -eq 1 ]; then
+    # CCSDS packing
+    # --------------
+    ${test_dir}/grib_encode_pthreads $ECCODES_SAMPLES_PATH/ccsds_grib2.tmpl
+    for f in temp.$label.out_*.grib; do
+        ${tools_dir}/grib_get -p min,max,avg $f > $f.ccsds.txt
+        diff $f.simple.txt $f.ccsds.txt
+        rm $f.simple.txt $f.ccsds.txt
+    done
+fi
 
-# CCSDS packing
-# --------------
-${test_dir}/grib_encode_pthreads $ECCODES_SAMPLES_PATH/ccsds_grib2.tmpl
+# Clean up
 for f in temp.$label.out_*.grib; do
-    ${tools_dir}/grib_get -p min,max,avg $f > $f.ccsds.txt
-    diff $f.simple.txt $f.ccsds.txt
-    rm $f.simple.txt $f.ccsds.txt
+    rm -f $f.simple.txt $f.ccsds.txt temp.$label.out_*.grib
 done
-
-rm -f temp.$label.out_*.grib
